requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==4.2.11
django-cors-headers==4.3.1
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
